requirements.txt
alembic==1.4.3
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-decouple==3.3
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
